Move ResultSphere to its own file
